+++ test/client/graphql/queries/reducer.test.tsx
@@ -148,4 +148,72 @@ describe('[queries] reducer', () => {
       </ApolloProvider>,
     );
   });
+
+  it('passes the prior props to the result-props mapper', done => {
+    const query: DocumentNode = gql`
+      query thing {
+        getThing {
+          thing
+        }
+      }
+    `;
+    const expectedData = { getThing: { thing: true } };
+    const link = mockSingleLink({
+      request: { query },
+      result: { data: expectedData },
+    });
+    const client = new ApolloClient({
+      link,
+      cache: new Cache({ addTypename: false }),
+    });
+
+    interface Data {
+      getThing: { thing: boolean };
+    }
+
+    interface FinalProps {
+      wrapper: { thingy: { thing: boolean } },
+      refetch: () => any,
+    }
+
+    const withData = graphql<{}, Data, {}, FinalProps>(query, {
+      props: ({ data }, lastProps: FinalProps) => {
+        const refetch = data!.refetch!;
+        let wrapper = { thingy: data!.getThing! };
+
+        // If the current thingy is equal to the last thingy,
+        // reuse the wrapper (to preserve referential equality).
+        if (lastProps && lastProps.wrapper.thingy === wrapper.thingy) {
+          wrapper = lastProps!.wrapper!;
+        }
+
+        return { wrapper, refetch };
+      },
+    });
+
+    let counter = 0;
+    class Container extends React.Component<FinalProps> {
+      componentWillReceiveProps(nextProps: FinalProps) {
+        expect(stripSymbols(nextProps.wrapper.thingy)).toEqual(expectedData.getThing);
+        if (counter === 1) {
+          expect(nextProps.wrapper).toEqual(this.props.wrapper);
+          done();
+        } else {
+          counter++;
+          nextProps.refetch();
+        }
+      }
+      render() {
+        return null;
+      }
+    }
+
+    const ContainerWithData = withData(Container);
+
+    renderer.create(
+      <ApolloProvider client={client}>
+        <ContainerWithData />
+      </ApolloProvider>,
+    );
+  });
 });